Fix kernel-devel dependency instalation with the kernel module
DKMS has dependency on the kernel-devel, but unfortunately
kernel-debug-devel is installed instead on Fedora/CentOS.
Changed OrderWithRequires statement to the Requires to ensure
kernel-devel is installed.
See more
here https://github.com/assuriosw/assurio-snap/issues/12 and
here https://bugzilla.redhat.com/1228897
This change ensures installation of the kernel-devel, not
kernel-debug-devel if both dkms and kernel module are installed
in one transaction. The installation order is preserved.

Left as it is for SLES.