Use versioned Vagrant boxes for CI
Now we have own workaround with usage of the Vagrant::Action::Builtin::BoxAdd triggers
The workaround is for the problem, when 2 parallel "vagrant up"s add or update a box.
Vagrant is using the same temporary file for download for both parallel downloads.
As result, file was corrupted and both or at list one "vagrant up" was failing.

NOTE: The version number is not strict. It's equal or hier than the defined
value. That's because, mainly, boxes are updated for elastio builds, not
elastio-snap builds. That's why the box version will be updated in this
repo less frequently than the new boxes will appear.