Fix detectin of the source branch
It became not a one line, that's desided to extract it to a script and
call it where it's necessary in different jobs.
Than decided to remove all bash-like GHA scripts to separate files from
the ci.yml

Also fixed wrong export of some branch names to a Vagrant box.