Add feature test for fn blk_alloc_queue changed in kernel 5.7
Add feature tests for the function `blk_alloc_queue`, which is changed in the kernel 5.7.
And, at the same moment was removed function `blk_queue_make_request` because it's
not necessary with the new version of the `blk_alloc_queue`.

**Note!** This change fixes the driver build with the 5.8 kernel, but doesn't fix work of the
driver with this kernel. There is an issue with the `make_request_fn` of the queue,
which is NULL.

Related to #48