Fix incremental snapshots on CentOS 7.4+
There was a fix for the 4.9 kernel and Debian 9 0035e38
It's great and in addition somehow fixes a problem of the driver and
loop devices. It relies on the 'req_op' structure presence, which
appeared in the kernel 4.9 and replaced with the 'req_opf' in the next
kernel versions.
CentOS 7 has enum req_op starting from the version 7.4, kernel
3.10.0-693, because they are doing backports from the 4th kernels.
But this enum has just 4 values instead of 6 as in other kernels,
where this enum is present. And it doesn't have defined REQ_OP_BITS,
which could could be defined and equal to the 2 bits.
As a fix, the condition is changed when to use passthrough 28 (kernel
4.9) instead of 30 (other kernels and CentOS's kernel 3.10).

Resolves #64