Linux kernel 5.8 support
A block device requests queue has a function 'make_request_fn' which is
used in the driver to send requests to the device underlyng to a
snapshot. Suddenly, starting from the kernel 5.8, this function has
become empty (NULL) in most cases.
Used call of the similar function 'blk_mq_make_request' in case if the
'make_request_fn' is not set. This call is wrapped into the
elastio_snap_null_mrf function with the memory barrier before call
blk_mq_make_request. Otherwise driver hangs on the snapshot destry
operation.

Resolves #44

Note: build of the driver on the Linux kernel 5.8 has been already fixed
in the #48, commit 96d63d.