Support Linux kernel 5.9 and 5.10
Fix thaw_bdev_int compat on Linux 5.9
- bd_private is suddenly gone from the block_device structure

Add 3 new compats for the Linux 5.9
- make_request_fn is gone from the requet_queue
- submit_bio is came to the block_device_operations
- blk_mq_submit_bio for kernels 5.9.0 - 5.9.1

Replace usage of the make_request_fn with the block_device_operations->sbmit_bio

Previously this driver was tracking changes by replacement of the
make_request_fn in the requests queue. Now in the Linux version 5.9 this
function has been replaced by the submit_bio function in the
block_device_operations structure. Unfortunately it's a const. So, there
is no possibility to replace it for the tracking purposes. As result,
now we are replacing entire block_device_operations structure in the
gendisk structure.

replase bdo->submit_bio with the blk_mq_submit_bio

Add symbol test for blk_mq_submit_bio in the kernel 5.9+

Fixes for kernels older than 5.9

Enable tests on Fedora 32