Fix tests on the loop dev on Linux 5.8
Fix, part 1

There was a failure like this:
`percpu ref (blk_queue_usage_counter_release) <= 0 (-149) after switching to atomic`
at the moment when test was trying to destroy a loop dev.

The queue's usage counter was not incremented each time when blk_mq_make_request
has been called. Fixed it by the manual increment, so it have to be >= 0 before
it's destroyed.
Interesting nuance that the issue is observed just on the loop devices,
but it seems to be not a loop dev specific.

Fix, part 2

elastio_snap_null_mrf function was mistakenly set to the original device after a
snapshot has been destroyed and tracking has been finished. Now it's set
back to the NULL.