Modify tests to cover ext2/3/4, xfs filesystems
This modification for the tests was done long time ago in scope of draft for #63
It allows to run `tests/elio-test.sh` with the file system specified as an argument.
Now tests can be executed like this
```bash
sudo ./elio-test.sh

sudo ./elio-test.sh -f ext3

sudo ./elio-test.sh -d /dev/vdb1 -f xfs
```

The test `test_track_writes` is disabled for XFS due to #63.
And all the tests for XFS were disabled on Debian 8 due to some mount issue
which is a part of issue #63 too.