Fix xfs v.4 fs consistency in snapshots
The idea of the fix to freeze fs instead of the freeze super.
The main idea was to repeat fix for the xfs v.3, where fs is frozen
before the freeze bdev. But in this case driver stuck time to time at
the not first freeze fs call. And, as tests demostrate, freeze fs is
enough for the fs-consistent snapshot