2nd draft: bring back sd_orig_mrf in the snap struct and add sd_tracing_ops
This change helps to avoid hack with the replacemnet of the const member
function `submit_bio` in the `block_device_operations` structure.
Also we are replacing entire struct for the traced device.
As experiment showed, this struct can be shared between multiple
devices. That was a root couse of the bug.

Also introduced new func `find_orig_fops` which is very like the
existing func `find_orig_mrf`. This is done to reduce count of
precompilation directives in the original fuction, which became
ureadable at all if still combined for kernels older and newer than 5.9.