! DO NOT SQUASH !
This patch handles the bug with multiple partitions observed in v5.16:

https://github.com/elastio/elastio-snap/issues/106

The issue was reproduced with ./elio-test.sh with additional loop device
partitions. No reproduces were observed for physical partitions neither
for real nor for the virtual block devices.

Starting from 5.16, block devices seem to have unique message queues,
which make the kernel call tracing_mrf() for each block device
separately, f.e., for loop0p1. Because of that, the dev->sd_sect_off
math worked incorrectly, forcing the driver to read the wrong sectors.
This, in its turn, leads to COW file corruption.

The idea of the fix is to redirect bio requests to a physical device,
thus making them executed in the old way, with the math logic unchanged.