Fix driver stuck on ext4 raid1 volume on Debian 9
= TL;DR:
This is reflection of the fix 49826d1 done by @skypodolsky for the
LVM/RAID setups and driver stuck in the loop of the
`snap_handle_write_bio` function. This time the same fix is applied for
the read bios in the `snap_handle_read_bio` function.

= Long description of the problem:
The investigation was started from the hung of the `test_modify_origin`
unit-test on Debian 9 with Linux kernel 4.9 on volume `raid1` and `ext4`
fs.
The test contains next steps:
1. setup snapshot of the raid volume
2. do some writes on the original volume
3. os.sync()
4. mount snapshot device
...

The 4th operation led to a hang of the `mount` operation. Interesting
detail, the hang was not reproducible without `os.sync()` in the test or
on `xfs` fs (because `xfs` doesn't honour sync operation).
The `mount` operation performs reads from the device. And in our case we
had infinite iterations inside of the `bio_for_each_segment` -> `while`
loop in the `snap_handle_read_bio` function. And `snap_cow_thread` was
loading 1 core for 100%. Usage of the `bio_for_each_segment_all` allows
to iterate via all segments of the bio and fix this problem.

Another interesting moment, that potentially this bug may occur on the
different kernel versions in cases if read bio are split for multiple
segments. But the most favorable conditions for this turned out to be
precisely after the sync operation before reading (mount call) on
kernel 4.9 with raid1.

Resolves #146