Fix kernel panic on snapshot destroy for a partition
The kernel panic was happening in case if there are multiple snapshot
devices are for the multiple partitions of the same disk. The problem
was for the kernels 5.9+.

There is a logic for these kernels to replace `block_device_operations`
structure with the driver's tracing function instead of an original
`submit_bio`. This struct belongs to the disk, and it's shared between
partitions of the disk.

The issue was in the access to the freed memory after ours tracing
struct was freed.

Now we are not allocating a new struct when setting up a snapshot for
the 2nd+ partition of the disk. And this struct is freed just when the
last snapshot for some partition of the disk has been destroyed.

The driver has the same behavior with the `make_request` function in
the bio queue for the kernel versions before 5.9. It's replaced with the
tracking function on the first setup snapshot operation for multiple
partitions of the disk. And an original `make_request` is set back when
the last snapshot device has been destroyed respectively.
So, now this behaviour is consistent for all Linux kernel versions.

Fixes #155