Implement functionality to switch on/of EIO propagation on snap read
Added an optional `-i` argument to `elioctl` to switch off IO errors
propagation on snapshot read when the snap is in the failed state.
The same arg has been added to the library functions to setup a snap and
transition to a snap.
As result, the entire error handling has been refactored a bit.
A've got rid of `sd_memory_fail_code` and `sd_cow_fail_state` snapshot
struct members. The error managing with usage of three error states, two
mentioned plus `sd_fail_code` was superfluous. Moreover, in case of
the snap dev memory mapping, IO errors were not propagated just in case
of the 2 errors: mem is too low and COW file size exceeded. Now all
other possible errors should not kill userspace app with `SIGBUS` if it
reads the snap as a memory-mapped file.