Downgrade kernels on Fedora 35/36
The new kernel 6 has been released and it's used on Fedora 37. And
Fedora 35 and 36 got updates with the new version.

However we do not support yet kernel 6 and Fedora 37.

Thus we need to downgrade kernels on these Fedora versions to fix CI.
It's also still useful to run CI tests on different kernel versions.