[Draft] Implement direct read/write IO (#223)
Implemented basic support of the direct IO functionality.

From now on, all read/write operations are sent directly to the device,
without any requirements for it to be mounted. To support this, we use
the fiemap ioctl called from the driver, to obtain all cow file extents. By
using these extents, we derive the needed sector for read/write operation.

Closes #220
Closes #221
Closes #224
Closes #225

---------

Co-authored-by: Alex Sereda <asereda@elastio.com>
Co-authored-by: Stanislav Barantsev <sbarantsev@elastio.com>