Fiemap doesn't work with XFS
It was found that in the implementation of fiemap in the xfs
driver there is a place that restricts 'length' by the 'long
long' type. Having used ~0ULL previously, which is actually
u64, caused int64 overflow on 'long long'. Replacing to INT_MAX
should be valid for the cow files so far.

Closes #227