Fix test_track_writes failure (#278)
During the test, we sometimes write to the sector already being tracked, which
results in a non-incrementing nr_changed_blocks counter. That happened because
a small amount of data was written (about 15 bytes) and the chances of writing to the
tracked sector were relatively high. The fix is to write more to increase the probability
of writing to the sector which isn't tracked.

Closes #167