Fix Makefile to call genconfig.sh only once
If called make -B after make clean, the behavior of Makefile was a bit inaccurate as we didn't
track the kernel-config.h dependency for the driver target. This caused the driver target to
create kernel-config.h underneath and also forced the root Makefile to regenerate it as a "fresh
one". Hence, the script was called twice, which took twice more time.

Closes #280