Implement snapshot/incremental reload functionality (#293)
The idea of this patch is to create reload scripts when the snapshot device is
created or changes its state, allowing the kernel to restore its state if the machine
has been restarted. These scripts are called on system start-up before the root
volume is mounted to ensure the snapshot device is ready and waiting for the
mount. When this occurs, the driver detects it, switches the device to an active
mode, and restores device tracking. The scripts are stored in the /etc/elastio/dla
directory.

Closes #289