Reload doesn't work on Ubuntu & Fedora (#296)
This patch introduces two fixes for Fedora and Ubuntu distributives to make the reload functionality work.

In general, this resolves the issues with false umounts. For Fedora, it was seen that the umount of the '.' directory
took place during operation. There were no results in finding the exact instance that does it and why, but it doesn't
look okay for regular operation. Hence, the code was corrected to accept only absolute paths instead of absolute +
relative (as in this situation).

For Ubuntu the problem was different. Additional mount points were created for the root volume, which caused
false umount handling in the driver. To fix that, we now check the vfsmount pointers to ensure we are dealing with
the root volume we are tracking initially.

Closes #292